reworked file structure, new device directory and split off framebuffer limine request

+++ kernel/src/device/tty.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-void init_tty(uint64_t cx, uint64_t cy, uint64_t w, uint64_t h);
+void init_tty(uint32_t cx, uint32_t cy, uint32_t w, uint32_t h);
 void printchar(char c);
 void printstr(const char* str);
 void printint(int num, int base);
